+++ homeassistant/components/intent/const.py
@@ -1,0 +1,3 @@
+"""Constants for the Intent integration."""
+
+DOMAIN = "intent"
