+++ homeassistant/components/intent/manifest.json
@@ -1,0 +1,11 @@
+{
+  "domain": "intent",
+  "name": "Intent",
+  "config_flow": false,
+  "documentation": "https://www.home-assistant.io/integrations/intent",
+  "requirements": [],
+  "ssdp": [],
+  "homekit": {},
+  "dependencies": ["http"],
+  "codeowners": ["@home-assistant/core"]
+}
